wip pull with submodules hopefully works now

+++ bin/wip/pull.sh
@@ -3,10 +3,16 @@
 set -e
 
 git pull origin wip -r
-git reset HEAD^
+if [[ $(git log -1 --pretty=%B) == 'wip' ]] ; then
+    git reset HEAD^
 
-git submodule foreach git pull origin wip -r
-git submodule foreach git reset HEAD^
-git submodule foreach git push origin --delete wip
+    git submodule foreach "
+        git pull origin wip -r
+        if [[ $(git log -1 --pretty=%B) == 'wip' ]] ; then
+            git reset HEAD^
+            git push origin --delete wip
+        fi
+    "
 
-git push origin --delete wip
+    git push origin --delete wip
+fi
